simplify update_stream_metas
