+++ src/query/service/src/catalogs/default/database_catalog.rs
@@ -95,9 +95,6 @@ use databend_common_meta_app::schema::UpdateIndexReply;
 use databend_common_meta_app::schema::UpdateIndexReq;
 use databend_common_meta_app::schema::UpdateMultiTableMetaReq;
 use databend_common_meta_app::schema::UpdateMultiTableMetaResult;
-use databend_common_meta_app::schema::UpdateStreamMetaReq;
-use databend_common_meta_app::schema::UpdateTableMetaReply;
-use databend_common_meta_app::schema::UpdateTableMetaReq;
 use databend_common_meta_app::schema::UpdateVirtualColumnReply;
 use databend_common_meta_app::schema::UpdateVirtualColumnReq;
 use databend_common_meta_app::schema::UpsertTableOptionReply;
@@ -570,26 +567,6 @@ impl Catalog for DatabaseCatalog {
     ) -> Result<UpsertTableOptionReply> {
         self.mutable_catalog
             .upsert_table_option(tenant, db_name, req)
-            .await
-    }
-
-    #[async_backtrace::framed]
-    async fn update_table_meta(
-        &self,
-        table_info: &TableInfo,
-        req: UpdateTableMetaReq,
-    ) -> Result<UpdateTableMetaReply> {
-        self.mutable_catalog
-            .update_table_meta(table_info, req)
-            .await
-    }
-
-    async fn update_stream_metas(
-        &self,
-        update_stream_meta_reqs: &[UpdateStreamMetaReq],
-    ) -> Result<()> {
-        self.mutable_catalog
-            .update_stream_metas(update_stream_meta_reqs)
             .await
     }
